0.1.4

+++ src/index.ts
@@ -16,6 +16,7 @@ export interface Config {
     init: boolean
     gain: number
     currency: string
+    divorce: boolean
 }
 
 export const Config: Schema<Config> = Schema.intersect([
@@ -24,6 +25,7 @@ export const Config: Schema<Config> = Schema.intersect([
           Schema.const('A').description('方案1，发送混合消息'),
           Schema.const('B').description('方案2，将消息分开发送'),
         ]).role('radio').default('A').deprecated().disabled(),
+        divorce: Schema.boolean().default(true).description('是否启用离婚功能').experimental(),
     }).description('基础配置'),
     Schema.object({
         database: Schema.boolean().default(true).description('是否启用数据库限制每日只能获取一个老婆'),
@@ -197,35 +199,37 @@ export async function apply(ctx: Context, cfg: Config) {
     })
 
     // 离婚指令
-    ctx.command('wife.离婚', '和你今天的老婆离婚').action(async ({session}) => {
-        if (session.onebot) {
-            if (session.subtype === 'group') {
-                if (cfg.database) {
-                    const groupId = session.channelId
-                    const userId = session.userId
-                    const uid = Number(userId)
-                    const get = await ctx.database.get('yuuzy_wife', {userId: userId, groupId: groupId})
-                    if (get.length > 0) {
-                        try {
-                            await ctx.monetary.cost(uid, 20, currency)
-                            await ctx.database.remove('yuuzy_wife', {userId: userId, groupId: groupId})
-                            await session.send(`离婚成功！\n你消耗了${gain}个货币。`)
-                        }catch (error) {
-                            await session.send('离婚失败！\n可能是因为你没有足够的货币。')
+    if (cfg.divorce) {
+        ctx.command('wife.离婚', '和你今天的老婆离婚').action(async ({session}) => {
+            if (session.onebot) {
+                if (session.subtype === 'group') {
+                    if (cfg.database) {
+                        const groupId = session.channelId
+                        const userId = session.userId
+                        const uid = Number(userId)
+                        const get = await ctx.database.get('yuuzy_wife', {userId: userId, groupId: groupId})
+                        if (get.length > 0) {
+                            try {
+                                await ctx.monetary.cost(uid, 20, currency)
+                                await ctx.database.remove('yuuzy_wife', {userId: userId, groupId: groupId})
+                                await session.send(`离婚成功！\n你消耗了${gain}个货币。`)
+                            }catch (error) {
+                                await session.send('离婚失败！\n可能是因为你没有足够的货币。')
+                            }
+                        }else {
+                            await session.send('你还没有老婆,快点娶一个吧！')
                         }
                     }else {
-                        await session.send('你还没有老婆,快点娶一个吧！')
+                        await session.send('数据库未启用。')
                     }
                 }else {
-                    await session.send('数据库未启用。')
+                 await session.send('请在群聊内使用！')
                 }
             }else {
-                await session.send('请在群聊内使用！')
+                await session.send('仅支持 OneBot 平台！')
             }
-        }else {
-            await session.send('仅支持 OneBot 平台！')
-        }
-    })
+        })
+    }
 
     // 定时清空数据库
     if (cfg.database) {
